Add default images and add image on category

+++ Services/FileService.cs
@@ -17,8 +17,11 @@ public class FileService
         return fileName;
     }
 
-    public static void DeleteFile(string? fileName)
+    public static void DeleteFile(string fileName)
     {
+        if (fileName.StartsWith("default_"))
+            return;
+        
         var path = Path.Combine(Directory.GetCurrentDirectory(), "Uploads", fileName);
         if (File.Exists(path))
             File.Delete(path);
